% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createGitHubRepositoryAndPush}
\alias{createGitHubRepositoryAndPush}
\title{Given a path to a directory that contians a local Git repository,
this function uses the GitHub web API to create a new repository under
the covidclinical organization, adds it as a remote to the local repository,
and pushes to the new GitHub remote.}
\usage{
createGitHubRepositoryAndPush(repositoryName, repositoryPath,
  private = FALSE)
}
\arguments{
\item{repositoryName}{The name of the repository}

\item{repositoryPath}{The path to the directory that contains the local git repository}
}
\description{
Given a path to a directory that contians a local Git repository,
this function uses the GitHub web API to create a new repository under
the covidclinical organization, adds it as a remote to the local repository,
and pushes to the new GitHub remote.
}
