% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createPhase2.1Stubs}
\alias{createPhase2.1Stubs}
\title{Creates R function stubs for a 4CE Project}
\usage{
createPhase2.1Stubs(projectName, workingDirectory,
  privateSummariesRepositoryName, publicSummariesRepositoryName)
}
\arguments{
\item{projectName}{The name of the project.}

\item{workingDirectory}{The directory in which the code stubs will be written.}

\item{privateSummariesRepositoryName}{The name of the repository for the private summary data for this project.}

\item{publicSummariesRepositoryName}{The name of the repository for the public summary data for this project.}
}
\description{
Creates R function stubs for a 4CE Project
}
