% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createPhase2Stubs}
\alias{createPhase2Stubs}
\title{Creates R function stubs for a 4CE Phase 2 Project}
\usage{
createPhase2Stubs(projectName, workingDirectory)
}
\arguments{
\item{projectName}{The name of the project.}

\item{workingDirectory}{The directory in which the code stubs will be written.}
}
\description{
Creates R function stubs for a 4CE Phase 2 Project
}
