% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{doInitializeAddCommit}
\alias{doInitializeAddCommit}
\title{Given a path to a directory, this function initialzies a new Git repository there,
then adds all of the current contents and does an initial commit.}
\usage{
doInitializeAddCommit(repositoryPath)
}
\arguments{
\item{repositoryPath}{The path to the directory that will become the git repository}
}
\description{
Given a path to a directory, this function initialzies a new Git repository there,
then adds all of the current contents and does an initial commit.
}
