% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{emit4CeFunctionStub}
\alias{emit4CeFunctionStub}
\title{Returns the text of the empty R function call stub for the project}
\usage{
emit4CeFunctionStub(projectName, functionName, commentPreamble,
  functionBody = "")
}
\arguments{
\item{projectName}{The name of the project.}

\item{functionName}{The name of the function to be emitted.}

\item{commentPreamble}{The preamble of the comments to be emitted.}

\item{functionBody}{Any text to be included in the function body.
", commentPreamble, " for the ", projectName, " project}
}
\description{
Returns the text of the empty R function call stub for the project
}
\keyword{4CE}
