% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{emit4CeFunctionStubToFile}
\alias{emit4CeFunctionStubToFile}
\title{Writes the result of calling emit4CeFunctionStub to a file named functionName.R}
\usage{
emit4CeFunctionStubToFile(projectName, functionName, commentPreamble,
  workingDirectory, functionBody = "", deparsedBody = FALSE)
}
\arguments{
\item{projectName}{The name of the project.}

\item{functionName}{The name of the function to be emitted.}

\item{commentPreamble}{The preamble of the comments to be emitted.}

\item{workingDirectory}{The working directory into which the file will be written}

\item{functionBody}{Any text to be included in the function body.}

\item{deparsedBody}{Was the text in functionBody the result of a call to deparse()?}
}
\description{
Writes the result of calling emit4CeFunctionStub to a file named functionName.R
}
