% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitHubCredentials.R
\name{getGitCredentials}
\alias{getGitCredentials}
\title{Call the git credential system to request the specified protocl / host credentials.
If not available, then ask the user to enter them manually.}
\usage{
getGitCredentials(protocol, host)
}
\arguments{
\item{protocol}{The protocol specification (e.g., "https") to be pased to the git credential helper}

\item{host}{The host specification (e.g., "github.com") to be pased to the git credential helper}
}
\description{
Call the git credential system to request the specified protocl / host credentials.
If not available, then ask the user to enter them manually.
}
\keyword{4CE}
