% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitHubCredentials.R
\name{interactiveSetGitCredentials}
\alias{interactiveSetGitCredentials}
\title{Ask user to enter username and password, let them review, then call 
setGitCredentials() to store it in the git credential manager.}
\usage{
interactiveSetGitCredentials(protocol, host)
}
\arguments{
\item{protocol}{The protocol specification (e.g., "https") to be pased to the git credential helper}

\item{host}{The host specification (e.g., "github.com") to be pased to the git credential helper}
}
\description{
Ask user to enter username and password, let them review, then call 
setGitCredentials() to store it in the git credential manager.
}
\keyword{4CE}
