% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitHubCredentials.R
\name{queryGitCredentialHelper}
\alias{queryGitCredentialHelper}
\title{Call the git credential system to request specified protocol / host credentials.
On success, it returns a named character vector with the requested credentials.
If the git credential helper fails to return the resutested credentials, it returns NA.}
\usage{
queryGitCredentialHelper(protocol, host)
}
\arguments{
\item{protocol}{The protocol specification (e.g., "https") to be pased to the git credential helper}

\item{host}{The host specification (e.g., "github.com") to be pased to the git credential helper}
}
\description{
Call the git credential system to request specified protocol / host credentials.
On success, it returns a named character vector with the requested credentials.
If the git credential helper fails to return the resutested credentials, it returns NA.
}
\keyword{4CE}
