% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitHubCredentials.R
\name{setGitCredentials}
\alias{setGitCredentials}
\title{Call the git credential helper to store credentials.}
\usage{
setGitCredentials(protocol, host, username, password)
}
\arguments{
\item{protocol}{The protocol specification (e.g., "https") to be pased to the git credential helper}

\item{host}{The host specification (e.g., "github.com") to be pased to the git credential helper}

\item{username}{The username to be stored}

\item{password}{The password to be stored}
}
\description{
Call the git credential helper to store credentials.
}
\keyword{4CE}
