% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{writeReadmeMd}
\alias{writeReadmeMd}
\title{Writes a placeholder README.md for a git repository}
\usage{
writeReadmeMd(projectName, workingDirectory, message = "")
}
\arguments{
\item{projectName}{The name of the git repository}

\item{workingDirectory}{The directory in which the README.md will be written.}
}
\description{
Writes a placeholder README.md for a git repository
}
