% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{Creates repository infrastructure for a 4CE Phase 2.1 Project.  Creates local repositories,
prompts for a GitHub username, and tries to create the corresponding GitHub repositories
and push inital contents.}
\usage{
createProject(projectName, workingDirectory = "/RDevelopment")
}
\arguments{
\item{projectName}{The name of the project.  Will be used to create correlated repository names.}

\item{workingDirectory}{The directory in which the project repositories will be created.}
}
\description{
Creates repository infrastructure for a 4CE Phase 2.1 Project.  Creates local repositories,
prompts for a GitHub username, and tries to create the corresponding GitHub repositories
and push inital contents.
}
\keyword{4CE}
